Move multi click safe components to shared module
