+++ collect_app/src/main/java/org/odk/collect/android/preferences/screens/MapsPreferencesFragment.kt
@@ -29,7 +29,7 @@ import org.odk.collect.android.preferences.CaptionedListPreference
 import org.odk.collect.android.preferences.PrefUtils
 import org.odk.collect.android.preferences.dialogs.ReferenceLayerPreferenceDialog
 import org.odk.collect.android.preferences.screens.ReferenceLayerPreferenceUtils.populateReferenceLayerPref
-import org.odk.collect.android.utilities.MultiClickGuard.allowClick
+import org.odk.collect.androidshared.ui.multiclicksafe.MultiClickGuard.allowClick
 import org.odk.collect.settings.keys.ProjectKeys.CATEGORY_BASEMAP
 import org.odk.collect.settings.keys.ProjectKeys.KEY_BASEMAP_SOURCE
 import java.io.File
