+++ collect_app/src/main/res/layout/date_widget_answer.xml
@@ -5,7 +5,7 @@
     android:paddingTop="@dimen/margin_small"
     android:paddingHorizontal="@dimen/margin_standard">
 
-    <org.odk.collect.android.views.multiclicksafe.MultiClickSafeButton
+    <org.odk.collect.androidshared.ui.multiclicksafe.MultiClickSafeButton
         android:id="@+id/date_button"
         style="@style/Widget.Collect.Button.WidgetAnswer"
         android:text="@string/select_date" />
