+++ collect_app/src/main/res/layout/manual_project_creator_dialog_layout.xml
@@ -161,7 +161,7 @@
         android:src="@drawable/shadow_up"
         app:layout_constraintBottom_toTopOf="@+id/cancel_button" />
 
-    <org.odk.collect.android.views.multiclicksafe.MultiClickSafeButton
+    <org.odk.collect.androidshared.ui.multiclicksafe.MultiClickSafeButton
         android:id="@+id/cancel_button"
         style="@style/Widget.MaterialComponents.Button.TextButton"
         android:layout_width="wrap_content"
@@ -171,7 +171,7 @@
         app:layout_constraintEnd_toStartOf="@+id/add_button"
         app:layout_constraintTop_toTopOf="@+id/add_button" />
 
-    <org.odk.collect.android.views.multiclicksafe.MultiClickSafeButton
+    <org.odk.collect.androidshared.ui.multiclicksafe.MultiClickSafeButton
         android:id="@+id/add_button"
         style="@style/Widget.MaterialComponents.Button.TextButton"
         android:layout_width="wrap_content"
