+++ collect_app/src/main/res/layout/select_minimal_widget_answer.xml
@@ -5,7 +5,7 @@
     android:paddingHorizontal="@dimen/margin_standard"
     android:paddingVertical="@dimen/margin_small">
 
-    <org.odk.collect.android.views.multiclicksafe.MultiClickSafeTextInputEditText
+    <org.odk.collect.androidshared.ui.multiclicksafe.MultiClickSafeTextInputEditText
         android:id="@+id/answer"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
